chore: move settings out of shared folder
