+++ app/common/renderer/actions/Inspector.js
@@ -1,7 +1,6 @@
 import _ from 'lodash';
 import {v4 as UUID} from 'uuid';
 
-import {SAVED_FRAMEWORK, SET_SAVED_GESTURES, getSetting, setSetting} from '../../shared/settings';
 import {APP_MODE, NATIVE_APP} from '../constants/session-inspector';
 import i18n from '../i18next';
 import AppiumClient from '../lib/appium-client';
@@ -15,6 +14,7 @@ import {
 } from '../utils/source-parsing';
 import {log} from '../utils/logger';
 import {showError} from './Session';
+import {SAVED_FRAMEWORK, SET_SAVED_GESTURES, getSetting, setSetting} from './settings';
 
 export const SET_SESSION_DETAILS = 'SET_SESSION_DETAILS';
 export const SET_SOURCE_AND_SCREENSHOT = 'SET_SOURCE_AND_SCREENSHOT';
