+++ app/common/renderer/actions/Session.js
@@ -14,14 +14,6 @@ import moment from 'moment';
 import {v4 as UUID} from 'uuid';
 import {Web2Driver} from 'web2driver';
 
-import {
-  SAVED_SESSIONS,
-  SERVER_ARGS,
-  SESSION_SERVER_PARAMS,
-  SESSION_SERVER_TYPE,
-  getSetting,
-  setSetting,
-} from '../../shared/settings';
 import {APP_MODE} from '../constants/session-inspector';
 import CloudProviders from '../components/Session/CloudProviders.jsx';
 import i18n from '../i18next';
@@ -29,6 +21,14 @@ import {fs, ipcRenderer, util} from '../polyfills';
 import {log} from '../utils/logger';
 import {addVendorPrefixes} from '../utils/other';
 import {quitSession, setSessionDetails} from './Inspector';
+import {
+  SAVED_SESSIONS,
+  SERVER_ARGS,
+  SESSION_SERVER_PARAMS,
+  SESSION_SERVER_TYPE,
+  getSetting,
+  setSetting,
+} from './settings';
 
 export const NEW_SESSION_REQUESTED = 'NEW_SESSION_REQUESTED';
 export const NEW_SESSION_LOADING = 'NEW_SESSION_LOADING';
